fix cookie repo and test it
