+++ tests/test_linkedin_client.py
@@ -1,7 +1,15 @@
 import pytest
 import json
 import os
+import sys
 from linkedin_api.client import Client
+
+TEST_LINKEDIN_USERNAME = os.getenv("LINKEDIN_USERNAME")
+TEST_LINKEDIN_PASSWORD = os.getenv("LINKEDIN_PASSWORD")
+
+if not (TEST_LINKEDIN_USERNAME and TEST_LINKEDIN_PASSWORD):
+    print("Test config incomplete. Exiting...")
+    sys.exit()
 
 
 @pytest.fixture
@@ -10,8 +18,6 @@ def client():
 
 
 def test_authenticate(client):
-    client.authenticate(
-        os.environ["LINKEDIN_USERNAME"], os.environ["LINKEDIN_PASSWORD"]
-    )
+    client.authenticate(TEST_LINKEDIN_USERNAME, TEST_LINKEDIN_USERNAME)
 
     assert client.session.cookies
